requirements.txt
streamlit==1.28.0
scikit-learn==1.3.0
python-dotenv==1.0.0
google-genai==0.3.0
fpdf==1.7.2